pyproject.toml
[tool.poetry]
name = "DirtyCam"
version = "0.1.0"
description = "Dirty cam detector"
authors = ["nnm"]
readme = "README.md"
packages = [{from = "src", include = "*"}]

[tool.poetry.dependencies]
python = "^3.10, <3.13"
opencv-python = "^4.11.0.86"
numpy  = "^1.26.4"
pandas = "^2.2.3"
scikit-learn = "^1.6.1"
catboost = "^1.2.7"
joblib = "^1.5.0"
pybrisque = "^1.0"
libsvm = {git = "https://github.com/akbargumbira/libsvm-python.git"}

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
ipykernel = "^6.29.5"

[tool.sort.config]
auto = true
case-sensitive = false
sort-python = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
extend-ignore = ["S101"]
line-length = 120
indent-width = 4
# See https://docs.astral.sh/ruff/settings/#target-version for more info.
target-version = "py310"

[tool.ruff.lint]
# Enable McCabe complexity ("C901"), errors ("E"), Pyflakes ("F"), isort ("I"),
# flake8-bandit ('S'), pycodestyle warnings (`W`)
select = ["C901", "E", "F", "I", "S", "W"]
ignore = []
# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"**/tests/*.py" = ["S"]

[tool.ruff.lint.isort]
# Until https://github.com/astral-sh/ruff/issues/2600 closed:
# multi_line_output = 5
# See https://docs.astral.sh/ruff/settings/#lintisort for more info.
known-third-party = ["alembic"]
known-first-party = ["config"]

[tool.ruff.format]
# Like Black:
# Use double quotes for strings.
# Indent with spaces, rather than tabs.
# Respect magic trailing commas.
# Automatically detect the appropriate line ending.
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.bandit]
skips = ["B101", "B106", "B308", "B311", "B404", "B603", "B608", "B703"]
